[#87] Setting: Security Config 설정 변경 및 토큰 만료 시간 변경
