+++ src/main/java/umc/GrowIT/Server/domain/Term.java
@@ -18,6 +18,7 @@ public class Term extends BaseEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    @Setter
     @Column(nullable = false, length = 50)
     private String title;
