+++ src/main/java/umc/GrowIT/Server/repository/UserChallengeRepository.java
@@ -1,7 +1,10 @@
 package umc.GrowIT.Server.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import umc.GrowIT.Server.domain.UserChallenge;
+
+import java.util.List;
 import java.util.Optional;
 
 public interface UserChallengeRepository extends JpaRepository<UserChallenge, Long> {
