--- src/main/java/umc/GrowIT/Server/web/controller/KakaoController.java
@@ -1,25 +1,0 @@
-package umc.GrowIT.Server.web.controller;
-
-import io.swagger.v3.oas.annotations.tags.Tag;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-import umc.GrowIT.Server.apiPayload.ApiResponse;
-import umc.GrowIT.Server.web.controller.specification.KakaoSpecification;
-import umc.GrowIT.Server.web.dto.UserDTO.KakaoResponseDTO;
-import umc.GrowIT.Server.service.authService.KakaoService;
-@Tag(name = "Kakao", description = "카카오 관련 API")
-
-@RestController
-@RequiredArgsConstructor
-public class KakaoController implements KakaoSpecification {
-
-    private final KakaoService kakaoService;
-
-    @GetMapping("/login/kakao")
-    public ApiResponse<KakaoResponseDTO.KakaoTokenDTO> kakaoLogin(@RequestParam(value = "code", required = false) String code) {
-        KakaoResponseDTO.KakaoTokenDTO token = kakaoService.getToken(code);
-        return ApiResponse.onSuccess(token);
-    }
-}
