--- src/main/java/umc/GrowIT/Server/web/controller/specification/KakaoSpecification.java
@@ -1,16 +1,0 @@
-package umc.GrowIT.Server.web.controller.specification;
-
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import umc.GrowIT.Server.apiPayload.ApiResponse;
-import umc.GrowIT.Server.web.dto.UserDTO.KakaoResponseDTO;
-
-public interface KakaoSpecification {
-
-    @GetMapping("/login/kakao")
-    @Operation(summary = "카카오 소셜 로그인", description = "", security = @SecurityRequirement(name = ""))
-    @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "COMMON200", description = "⭕ SUCCESS")
-    ApiResponse<KakaoResponseDTO.KakaoTokenDTO> kakaoLogin(@RequestParam(value = "code", required = false) String code);
-}
