+++ src/main/java/umc/GrowIT/Server/web/dto/UserDTO/UserRequestDTO.java
@@ -1,9 +1,6 @@
 package umc.GrowIT.Server.web.dto.UserDTO;
 
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotEmpty;
-import jakarta.validation.constraints.Size;
+import jakarta.validation.constraints.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
